Remove support of neutron-lbaas from lb_listener (#1695)
