+++ openstack/lb_v2_shared.go
@@ -10,7 +10,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
 	"github.com/gophercloud/gophercloud"
-	octavialisteners "github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/listeners"
 	octavialoadbalancers "github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/loadbalancers"
 	octaviapools "github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/pools"
 	neutronl7policies "github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas_v2/l7policies"
@@ -131,301 +130,6 @@ func chooseLBV2LoadbalancerUpdateOpts(d *schema.ResourceData, config *Config) (n
 	}
 
 	return nil, nil
-}
-
-// chooseLBV2ListenerCreateOpts will determine which load balancer listener Create options to use:
-// either the Octavia/LBaaS or the Neutron/Networking v2.
-func chooseLBV2ListenerCreateOpts(d *schema.ResourceData, config *Config) (neutronlisteners.CreateOptsBuilder, error) {
-	adminStateUp := d.Get("admin_state_up").(bool)
-
-	var sniContainerRefs []string
-	if raw, ok := d.GetOk("sni_container_refs"); ok {
-		for _, v := range raw.([]interface{}) {
-			sniContainerRefs = append(sniContainerRefs, v.(string))
-		}
-	}
-
-	var createOpts neutronlisteners.CreateOptsBuilder
-
-	if config.UseOctavia {
-		// Use Octavia.
-		opts := octavialisteners.CreateOpts{
-			// Protocol SCTP requires octavia minor version 2.23
-			Protocol:               octavialisteners.Protocol(d.Get("protocol").(string)),
-			ProtocolPort:           d.Get("protocol_port").(int),
-			ProjectID:              d.Get("tenant_id").(string),
-			LoadbalancerID:         d.Get("loadbalancer_id").(string),
-			Name:                   d.Get("name").(string),
-			DefaultPoolID:          d.Get("default_pool_id").(string),
-			Description:            d.Get("description").(string),
-			DefaultTlsContainerRef: d.Get("default_tls_container_ref").(string),
-			SniContainerRefs:       sniContainerRefs,
-			AdminStateUp:           &adminStateUp,
-		}
-
-		if v, ok := d.GetOk("connection_limit"); ok {
-			connectionLimit := v.(int)
-			opts.ConnLimit = &connectionLimit
-		}
-
-		if v, ok := d.GetOk("timeout_client_data"); ok {
-			timeoutClientData := v.(int)
-			opts.TimeoutClientData = &timeoutClientData
-		}
-
-		if v, ok := d.GetOk("timeout_member_connect"); ok {
-			timeoutMemberConnect := v.(int)
-			opts.TimeoutMemberConnect = &timeoutMemberConnect
-		}
-
-		if v, ok := d.GetOk("timeout_member_data"); ok {
-			timeoutMemberData := v.(int)
-			opts.TimeoutMemberData = &timeoutMemberData
-		}
-
-		if v, ok := d.GetOk("timeout_tcp_inspect"); ok {
-			timeoutTCPInspect := v.(int)
-			opts.TimeoutTCPInspect = &timeoutTCPInspect
-		}
-
-		if v, ok := d.GetOk("tags"); ok {
-			tags := v.(*schema.Set).List()
-			opts.Tags = expandToStringSlice(tags)
-		}
-
-		// Get and check insert  headers map.
-		rawHeaders := d.Get("insert_headers").(map[string]interface{})
-		headers, err := expandLBV2ListenerHeadersMap(rawHeaders)
-		if err != nil {
-			return nil, fmt.Errorf("unable to parse insert_headers argument: %s", err)
-		}
-
-		opts.InsertHeaders = headers
-
-		if raw, ok := d.GetOk("allowed_cidrs"); ok {
-			allowedCidrs := make([]string, len(raw.([]interface{})))
-			for i, v := range raw.([]interface{}) {
-				allowedCidrs[i] = v.(string)
-			}
-			opts.AllowedCIDRs = allowedCidrs
-		}
-
-		createOpts = opts
-
-		return createOpts, nil
-	}
-
-	// Use Neutron.
-	opts := neutronlisteners.CreateOpts{
-		Protocol:               neutronlisteners.Protocol(d.Get("protocol").(string)),
-		ProtocolPort:           d.Get("protocol_port").(int),
-		TenantID:               d.Get("tenant_id").(string),
-		LoadbalancerID:         d.Get("loadbalancer_id").(string),
-		Name:                   d.Get("name").(string),
-		DefaultPoolID:          d.Get("default_pool_id").(string),
-		Description:            d.Get("description").(string),
-		DefaultTlsContainerRef: d.Get("default_tls_container_ref").(string),
-		SniContainerRefs:       sniContainerRefs,
-		AdminStateUp:           &adminStateUp,
-	}
-
-	if v, ok := d.GetOk("connection_limit"); ok {
-		connectionLimit := v.(int)
-		opts.ConnLimit = &connectionLimit
-	}
-
-	createOpts = opts
-
-	return createOpts, nil
-}
-
-// chooseLBV2ListenerUpdateOpts will determine which load balancer listener Update options to use:
-// either the Octavia/LBaaS or the Neutron/Networking v2.
-func chooseLBV2ListenerUpdateOpts(d *schema.ResourceData, config *Config) (neutronlisteners.UpdateOptsBuilder, error) {
-	var hasChange bool
-
-	if config.UseOctavia {
-		// Use Octavia.
-		var opts octavialisteners.UpdateOpts
-		if d.HasChange("name") {
-			hasChange = true
-			name := d.Get("name").(string)
-			opts.Name = &name
-		}
-
-		if d.HasChange("description") {
-			hasChange = true
-			description := d.Get("description").(string)
-			opts.Description = &description
-		}
-
-		if d.HasChange("connection_limit") {
-			hasChange = true
-			connLimit := d.Get("connection_limit").(int)
-			opts.ConnLimit = &connLimit
-		}
-
-		if d.HasChange("timeout_client_data") {
-			hasChange = true
-			timeoutClientData := d.Get("timeout_client_data").(int)
-			opts.TimeoutClientData = &timeoutClientData
-		}
-
-		if d.HasChange("timeout_member_connect") {
-			hasChange = true
-			timeoutMemberConnect := d.Get("timeout_member_connect").(int)
-			opts.TimeoutMemberConnect = &timeoutMemberConnect
-		}
-
-		if d.HasChange("timeout_member_data") {
-			hasChange = true
-			timeoutMemberData := d.Get("timeout_member_data").(int)
-			opts.TimeoutMemberData = &timeoutMemberData
-		}
-
-		if d.HasChange("timeout_tcp_inspect") {
-			hasChange = true
-			timeoutTCPInspect := d.Get("timeout_tcp_inspect").(int)
-			opts.TimeoutTCPInspect = &timeoutTCPInspect
-		}
-
-		if d.HasChange("default_pool_id") {
-			hasChange = true
-			defaultPoolID := d.Get("default_pool_id").(string)
-			opts.DefaultPoolID = &defaultPoolID
-		}
-
-		if d.HasChange("default_tls_container_ref") {
-			hasChange = true
-			defaultTLSContainerRef := d.Get("default_tls_container_ref").(string)
-			opts.DefaultTlsContainerRef = &defaultTLSContainerRef
-		}
-
-		if d.HasChange("sni_container_refs") {
-			hasChange = true
-			var sniContainerRefs []string
-			if raw, ok := d.GetOk("sni_container_refs"); ok {
-				for _, v := range raw.([]interface{}) {
-					sniContainerRefs = append(sniContainerRefs, v.(string))
-				}
-			}
-			opts.SniContainerRefs = &sniContainerRefs
-		}
-
-		if d.HasChange("admin_state_up") {
-			hasChange = true
-			asu := d.Get("admin_state_up").(bool)
-			opts.AdminStateUp = &asu
-		}
-
-		if d.HasChange("insert_headers") {
-			hasChange = true
-
-			// Get and check insert headers map.
-			rawHeaders := d.Get("insert_headers").(map[string]interface{})
-			headers, err := expandLBV2ListenerHeadersMap(rawHeaders)
-			if err != nil {
-				return nil, fmt.Errorf("unable to parse insert_headers argument: %s", err)
-			}
-
-			opts.InsertHeaders = &headers
-		}
-
-		if d.HasChange("allowed_cidrs") {
-			hasChange = true
-			var allowedCidrs []string
-			if raw, ok := d.GetOk("allowed_cidrs"); ok {
-				for _, v := range raw.([]interface{}) {
-					allowedCidrs = append(allowedCidrs, v.(string))
-				}
-			}
-			opts.AllowedCIDRs = &allowedCidrs
-		}
-
-		if d.HasChange("tags") {
-			hasChange = true
-			if v, ok := d.GetOk("tags"); ok {
-				tags := v.(*schema.Set).List()
-				tagsToUpdate := expandToStringSlice(tags)
-				opts.Tags = &tagsToUpdate
-			} else {
-				opts.Tags = &[]string{}
-			}
-		}
-
-		if hasChange {
-			return opts, nil
-		}
-	}
-
-	// Use Neutron.
-	var opts neutronlisteners.UpdateOpts
-	if d.HasChange("name") {
-		hasChange = true
-		name := d.Get("name").(string)
-		opts.Name = &name
-	}
-
-	if d.HasChange("description") {
-		hasChange = true
-		description := d.Get("description").(string)
-		opts.Description = &description
-	}
-
-	if d.HasChange("connection_limit") {
-		hasChange = true
-		connLimit := d.Get("connection_limit").(int)
-		opts.ConnLimit = &connLimit
-	}
-
-	if d.HasChange("default_pool_id") {
-		hasChange = true
-		defaultPoolID := d.Get("default_pool_id").(string)
-		opts.DefaultPoolID = &defaultPoolID
-	}
-
-	if d.HasChange("default_tls_container_ref") {
-		hasChange = true
-		defaultTLSContainerRef := d.Get("default_tls_container_ref").(string)
-		opts.DefaultTlsContainerRef = &defaultTLSContainerRef
-	}
-
-	if d.HasChange("sni_container_refs") {
-		hasChange = true
-		var sniContainerRefs []string
-		if raw, ok := d.GetOk("sni_container_refs"); ok {
-			for _, v := range raw.([]interface{}) {
-				sniContainerRefs = append(sniContainerRefs, v.(string))
-			}
-		}
-		opts.SniContainerRefs = &sniContainerRefs
-	}
-
-	if d.HasChange("admin_state_up") {
-		hasChange = true
-		asu := d.Get("admin_state_up").(bool)
-		opts.AdminStateUp = &asu
-	}
-
-	if hasChange {
-		return opts, nil
-	}
-
-	return nil, nil
-}
-
-func expandLBV2ListenerHeadersMap(raw map[string]interface{}) (map[string]string, error) {
-	m := make(map[string]string, len(raw))
-	for key, val := range raw {
-		labelValue, ok := val.(string)
-		if !ok {
-			return nil, fmt.Errorf("label %s value should be string", key)
-		}
-
-		m[key] = labelValue
-	}
-
-	return m, nil
 }
 
 func waitForLBV2Listener(ctx context.Context, lbClient *gophercloud.ServiceClient, listener *neutronlisteners.Listener, target string, pending []string, timeout time.Duration) error {
